+++ py/desiutil/plots.py
@@ -7,18 +7,9 @@ desiutil.plots
 
 Module for code plots.
 """
-from __future__ import (print_function, absolute_import, division,
-                        unicode_literals)
-
 import os
-
 import numpy as np
 import numpy.ma
-
-try:
-    basestring
-except NameError:  # For Python 3
-    basestring = str
 
 
 def plot_slices(x, y, x_lo, x_hi, y_cut, num_slices=5, min_count=100, axis=None,
@@ -118,7 +109,6 @@ def plot_slices(x, y, x_lo, x_hi, y_cut, num_slices=5, min_count=100, axis=None,
     # ylim
     if set_ylim_from_stats:
         axis.set_ylim(min_m2-max_yr/2., max_p2+max_yr/2.)
-
 
     # Plot cut lines.
     axis.axhline(+y_cut, ls=':', color='k')
@@ -293,7 +283,7 @@ def prepare_data(data, mask=None, clip_lo=None, clip_hi=None,
     # Convert percentile clip values to absolute values.
     def get_clip(value):
         clip_mask = False
-        if isinstance(value, basestring):
+        if isinstance(value, str):
             if value.startswith('!'):
                 clip_mask = True
                 value = value[1:]
@@ -330,7 +320,7 @@ def prepare_data(data, mask=None, clip_lo=None, clip_hi=None,
 
 
 def init_sky(projection='eck4', ra_center=120, galactic_plane_color='red',
-             ra_labels=np.arange(0, 360, 60),
+             ecliptic_color='red', ra_labels=np.arange(0, 360, 60),
              dec_labels=np.arange(-60, 90, 30), ax=None):
     """Initialize a basemap projection of the full sky.
 
@@ -361,7 +351,10 @@ def init_sky(projection='eck4', ra_center=120, galactic_plane_color='red',
         Map is centered at this RA in degrees. Default is +120, which
         avoids splitting the DESI northern and southern regions.
     galactic_plane_color : color name or None
-        Draw a line representing the galactic plane using the specified
+        Draw a curve representing the galactic plane using the specified
+        color, or do nothing when None.
+    ecliptic_color : color name or None
+        Draw a dotted curve representing the ecliptic plane using the specified
         color, or do nothing when None.
     ra_labels : iterable or None
         List of RA values in degrees that will be labeled on the map.
@@ -384,7 +377,7 @@ def init_sky(projection='eck4', ra_center=120, galactic_plane_color='red',
     from matplotlib.patches import Polygon
     from mpl_toolkits.basemap import pyproj
     from mpl_toolkits.basemap import Basemap
-    from astropy.coordinates import SkyCoord
+    from astropy.coordinates import SkyCoord, HeliocentricTrueEcliptic, ICRS
     import astropy.units as u
 
     # Define a Basemap subclass with an ellipse() method.
@@ -488,6 +481,17 @@ def init_sky(projection='eck4', ra_center=120, galactic_plane_color='red',
             galactic_x, galactic_y, marker='.', s=20, lw=0, alpha=0.75,
             c=galactic_plane_color)
         # Make sure the galactic plane stays above other displayed objects.
+        paths.set_zorder(20)
+
+    # Draw the optional ecliptic plane.
+    if ecliptic_color is not None:
+        lon = np.linspace(0, 2 * np.pi, 50) * u.rad
+        ecliptic = HeliocentricTrueEcliptic(
+            lon=lon, lat=0 * u.radian, distance=1 * u.Mpc).transform_to(ICRS)
+        ecliptic_x, ecliptic_y = m(ecliptic.ra.degree, ecliptic.dec.degree)
+        paths = m.scatter(
+            ecliptic_x, ecliptic_y, marker='.', s=20, lw=0, alpha=0.75,
+            c=ecliptic_color)
         paths.set_zorder(20)
 
     return m