+++ py/desiutil/test/test_io.py
@@ -2,19 +2,14 @@
 # -*- coding: utf-8 -*-
 """Test desiutil.io.
 """
-from __future__ import (absolute_import, division,
-                        print_function, unicode_literals)
-# The line above will help with 2to3 support.
 import unittest
+import os
+import stat
 import sys
+from tempfile import TemporaryDirectory
 import numpy as np
 from astropy.table import Table
-from ..io import combine_dicts, decode_table, encode_table, yamlify
-
-try:
-    basestring
-except NameError:  # For Python 3
-    basestring = str
+from ..io import combine_dicts, decode_table, encode_table, yamlify, unlock_file
 
 
 class TestIO(unittest.TestCase):
@@ -128,17 +123,18 @@ class TestIO(unittest.TestCase):
                  'lst': ['tst2', np.int16(2)],
                  'tup': (1, 3), 'dct': {'a': 'tst3', 'b': np.float32(6.)},
                  'array': np.zeros(10)}
-        if sys.version_info >= (3, 0, 0):
-            self.assertIsInstance(fdict['name'], str)
-        else:
-            self.assertIsInstance(fdict['name'], unicode)
+        self.assertIsInstance(fdict['name'], str)
         # Run
         ydict = yamlify(fdict)
         self.assertIsInstance(ydict['flt32'], float)
         self.assertIsInstance(ydict['array'], list)
         for key in ydict.keys():
-            if isinstance(key, basestring):
+            if isinstance(key, str):
+                # This looks a little silly, but in fact, some of the keys
+                # are integers not strings.
                 self.assertIsInstance(key, str)
+            else:
+                self.assertIsInstance(key, int)
 
     def test_combinedicts(self):
         """Test combining dicts
@@ -177,6 +173,32 @@ class TestIO(unittest.TestCase):
         self.assertEqual(dict3, {'a': {'b': {'x': 1, 'y': 2, 'p': 3, 'q': 4}}})
         self.assertEqual(dict1, {'a': {'b': {'x': 1, 'y': 2}}})
         self.assertEqual(dict2, {'a': {'b': {'p': 3, 'q': 4}}})
+
+    def test_unlock_file(self):
+        """Test the permission unlock file manager.
+        """
+        fff = stat.S_IRUSR | stat.S_IRGRP | stat.S_IROTH
+        www = stat.S_IWUSR | stat.S_IWGRP | stat.S_IWOTH
+        with TemporaryDirectory() as dirname:
+            filename = os.path.join(dirname, 'tempfile')
+            with open(filename, 'wb') as f:
+                f.write(b'Content\n')
+            s0 = os.stat(filename)
+            ro = stat.S_IFMT(s0.st_mode) | fff
+            os.chmod(filename, ro)
+            s1 = os.stat(filename)
+            self.assertEqual(stat.S_IMODE(s1.st_mode), fff)
+            with unlock_file(filename, 'ab') as f:
+                f.write(b'More content\n')
+                s2 = os.stat(filename)
+                self.assertEqual(stat.S_IMODE(s2.st_mode), fff | stat.S_IWUSR)
+            s3 = os.stat(filename)
+            self.assertEqual(stat.S_IMODE(s3.st_mode), fff)
+            filename = os.path.join(dirname, 'newfile')
+            with unlock_file(filename, 'wb') as f:
+                f.write(b'Some content\n')
+            s0 = os.stat(filename)
+            self.assertEqual(stat.S_IMODE(s0.st_mode) & www, 0)
 
 
 def test_suite():